Ignore Copy File If Not Exists

+++ flare-download-observations/flare-container.sh
@@ -143,7 +143,7 @@ GIT_REMOTE_SSHKEYPRIVATE_FILE=$(awk -F/ '{print $NF}' <<< ${GIT_REMOTE_SSHKEYPRI
 
 # Setup SSH
 mkdir -p /root/.ssh
-cp -u ${DIRECTORY_CONTAINER_SHARED}/${CONTAINER_NAME}/${GIT_REMOTE_SSHKEYPRIVATE_FILE} /root/.ssh/id_rsa
+[[ -e ${DIRECTORY_CONTAINER_SHARED}/${CONTAINER_NAME}/${GIT_REMOTE_SSHKEYPRIVATE_FILE} ]] && cp -u ${DIRECTORY_CONTAINER_SHARED}/${CONTAINER_NAME}/${GIT_REMOTE_SSHKEYPRIVATE_FILE} /root/.ssh/id_rsa
 # TODO: Add All Remote Servers to known_hosts
 ssh-keyscan ${GIT_REMOTE_SERVER} > /root/.ssh/known_hosts
